# target/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Parent pom providing dependency and plugin management for applications\n\t\tbuilt with Maven",
    "version" : "1.0.0",
    "title" : "forum-main"
  },
...
  "paths" : {
...
    "/api/crud/user/user/users" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "User 列表信息",
        "description" : "",
        "operationId" : "queryUserPage",
        "parameters" : [ {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "format" : "date-time"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User对象"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "User" ],
        "summary" : "新建 User",
        "description" : "",
        "operationId" : "createUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User对象"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User对象"
            }
          }
        }
      }
    },
    "/api/crud/user/user/users/{id}" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "查看 User",
        "description" : "",
        "operationId" : "getUser",
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User对象"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "User" ],
        "summary" : "修改 User",
        "description" : "",
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User对象"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User对象"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "User" ],
        "summary" : "删除 User",
        "description" : "",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Tip"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Tip" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
    "User对象" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "用户ID"
        },
        "account" : {
          "type" : "string"
        },
        "uname" : {
          "type" : "string",
          "description" : "用户名"
        },
        "password" : {
          "type" : "string",
          "description" : "密码"
        },
        "email" : {
          "type" : "string",
          "description" : "邮箱"
        },
        "flag" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "权限"
        },
        "headUrl" : {
          "type" : "string",
          "description" : "头像地址"
        },
        "createTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        }
      }
    }
  }
}